LayoutTests/platform/ios/webaudio/audioprocessingevent-expected.txt

Tests AudioProcessingEvent attributes in IDL.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS event.playbackTime is defined.
PASS event.inputBuffer is defined.
PASS event.outputBuffer is defined.
PASS successfullyParsed is true

TEST COMPLETE
